refactor: Movendo o multer para fora das controllers
