+++ src/server/controllers/AccountPicture.ts
@@ -1,35 +1,7 @@
-import multer from 'multer'
 import { StatusCodes } from "http-status-codes";
 import { Request, Response } from "express";
 import path from 'path'
 import fs from 'fs'
-
-const storage = multer.diskStorage({
-    destination: process.env.PROFILE_DEST,
-    filename: (req, file, cb) => {
-        const extensionName = path.extname(file.originalname)        
-        cb(null,req.params.id + (extensionName == '.jpeg' ? '.jpg' : extensionName));
-    } 
-});
-
-export const upload = multer({
-    storage: storage,
-    limits: { fileSize: Number(process.env.PROFILE_MAXSIZE)},
-    fileFilter: function(_req, file, cb){
-        const allowedFileExtensions = ['.jpg', '.jpeg'];
-        const extension = path.extname(file.originalname).toLowerCase();
-        if (!allowedFileExtensions.includes(extension)) {
-            return cb(new Error('Only .jpg and .jpeg files are allowed'));
-        }
-
-        const allowedMimeTypes = ['image/jpeg', 'image/jpg'];
-        if (!allowedMimeTypes.includes(file.mimetype)) {
-            return cb(new Error('Only .jpeg and .jpg image mimetypes are allowed'));
-        }
-
-        cb(null, true);
-    },
-});
 
 const uploadAvatar = async (req: Request, res: Response) => {
     if (!req.file) {
@@ -85,7 +57,6 @@ const getAvatar = async(_req: Request, res: Response) => {
 }
 
 export const pictureHandler = {
-    upload,
     uploadAvatar,
     removeAvatar,
     getAvatar,
